+++ tech-far-hub/content/components/global-nav.yaml
